FiniteLattice

Latlib.FiniteLattice — Type

FiniteLattice

A lattice with a finite number of sites confined by a boundary box. The boundary box vectors$\mathcal{B}_i$ are specified as integer multiples of the lattice vectors. The cartesian boundary box vectors $\mathbf{b}_i$ are then given by,

$\mathbf{b}_i = \mathbf{A}\mathcal{B}_i$

Arguments

lattice::Lattice: instance of the underlying lattice geometry.

boundary::AbstractMatrix: $D \times D$ integer matrix whose columns define the boundary box

periodicity::Vector{Bool}: vector defining which boundary direction is periodic

order: functions setting

Latlib.FiniteLattice — Method

FiniteLattice(lattice::Lattice, boundary::AbstractMatrix; periodic=true)

Create a finite lattice given an underlying lattice, its boundary, and optionally whether or not fully periodic boundary conditions need to be applied

Arguments

lattice::Lattice: instance of the underlying lattice geometry.

boundary::AbstractMatrix: $D \times D$ integer matrix whose columns define the boundary box

Keyword arguments

periodic::Bool=true: flag to set if periodic boundary conditions are used

Latlib.coordinates — Method

coordinates(flattice::FiniteLattice)

Computes all coordinates of the finite lattice within the boundary box.

Latlib.bravais_coordinates — Method

bravais_coordinates(flattice::FiniteLattice)

Computes the Bravais coordinates of the finite lattice within the boundary box, i.e. only the coordinates of the Bravais lattice but not the coordinates of the atomic positions

Latlib.boundary_vectors — Method

boundary_vectors(flattice::FiniteLattice)

Computes the vectors of the boundary box in Cartesian coordinates

Latlib.periodicity_vectors — Method

periodicity_vectors(flattice::FiniteLattice)

Computes the vectors of the periodic directions in Cartesian coordinates

Latlib.neighbors — Method

neighbors(flattice::FiniteLattice; num_distance::Integer=1)

Computes which pairs of coordinates are neighbors

Arguments

flattice::FiniteLattice: finite lattice

Keyword arguments

num_distance::Integer=1: at which distance neighbors are considered, 1 -> nearest neighbor, 2 -> second nearest neighbor, etc.

Latlib.distance — Method

distance(x1::AbstractVector, x2::AbstractVector, flattice::FiniteLattice)

Computes the distance between two points given the periodicity of the finite lattice.

Latlib.distances — Method

distances(flattice::FiniteLattice)

Computes which unique values of distances are present on the lattice

Latlib.dimension — Method

dimension(flattice::FiniteLattice)

Obtain the dimension of the lattice.

Latlib.natoms — Method

natoms(flattice::FiniteLattice)

Obtain the number of atomic positions