Metric

Latlib.distance — Method

distance(x1::AbstractVector, x2::AbstractVector; periodicity_vectors=nothing)

Computes the distance between two points

Arguments

x1::AbstractVector: first point

x2::AbstractVector: second point

Keyword arguments

periodicity_vectors=nothing: if defined, the vectors that define a periodic direction

If periodicity_vectors is not defined, the euclidian distance is computed

$d(\mathbf{x}_1, \mathbf{x}_2) = \lVert \mathbf{x}_1 - \mathbf{x}_2 \rVert$.

The periodicity vectors are a matrix $\mathcal{P}= (\mathbf{p}_1 | \ldots | \mathbf{p}_p)$ of dimension $D \times p$ where $D$ denotes the dimension of the points, and $p$ denotes the number of periodic directions.

If periodicity_vectors is defined the distance is defined as,

$d(\mathbf{x}_1, \mathbf{x}_2; \mathcal{P}) \equiv \min\limits_{n_1, \ldots, n_p \in \mathbf{Z}} \lVert \mathbf{x}_1 + \sum_{i=1}^p n_i \mathbf{p}_i - \mathbf{x}_2 \rVert$

Latlib.distance_matrix — Method

distance_matrix(points::AbstractMatrix; periodicity_vectors=nothing)

Computes the pairwise distances between points

Arguments

points::AbstractMatrix: matrix whose columns are the points of which the distance is computed

Keyword arguments

periodicity_vectors=nothing: if defined, the vectors that define a periodic direction

Latlib.distances — Method

distances(points::AbstractMatrix; periodicity_vectors=nothing)

Computes which unique values of distances are present between the points

Arguments

points::AbstractMatrix: matrix whose columns are the points of which the distance is computed

Keyword arguments

periodicity_vectors=nothing: if defined, the vectors that define a periodic direction

Latlib.neighbors — Method

neighbors(points::AbstractMatrix; num_distance::Integer=1, periodicity_vectors=nothing)

Computes which pairs of the input points are neighbors

Arguments

points::AbstractMatrix: matrix whose columns are the points of which the distance is computed

Keyword arguments

num_distance::Integer=1: at which distance neighbors are considered, 1 -> nearest neighbor, 2 -> second nearest neighbor, etc.

periodicity_vectors=nothing: if defined, the vectors that define a periodic direction