OpSum

Latlib.Op — Type

Op

Defines the information regarding an operator $\mathbf{O}$.

Arguments

type::AbstractString: String specifing the type of operator: "HB", "Cup", etc ...

coupling::Union{AbstractString,Number}: String or number specifying the coupling constant;

sites::Vector{Int64}: Vector specifing the sites in each the operator acts on the Hilbert Space;

Latlib.OpSum — Type

OpSum

Constructor to collect all operators that appear in the Hamiltonian.

One can use the += operator to add operators to the OpSum:

    opsum = OpSum()
    opsum += Op("HB", "Jd", [1,2])
    opsum += Op("HB", "Jd", [2,3])
    opsum += Op("HB", "Jd", [3,4])

Latlib.neighbor_bonds — Method

Returns the OpSum() for corresponding to a two-body operator acting between two neighboring sites.

Arguments

type::AbstractString: String specifing the type of operator: "HB", "Cup", etc ...

coupling::AbstractString: String or number specifying the coupling constant;

lattice::FiniteLattice: The lattice on which the operator acts;

num_distance::Integer=1: at which distance neighbors are considered, 1 -> nearest neighbor, 2 -> second nearest neighbor, etc.

Latlib.lattice_bonds — Method

Returns the OpSum() for the corresponding two-body operator acting on the sites specified by the vectors b1 and b2.

The vectors b1 and b2 are given in the following way:
- b_[1] = index of the atom in the unit cell (using the same order as given to the lattice instance)
- b_[2] = Coordinate of the unit cell along the first Bravais vector
- b_[3] = Coordinate of the unit cell along the second Bravais vector
- b_[4] = Coordinate of the unit cell along the third Bravais vector
- ...

Arguments

type::AbstractString: String specifing the type of operator: "HB", "Cup", etc ...

coupling::AbstractString: String or number specifying the coupling constant;

lattice::FiniteLattice: The lattice on which the operator acts;

b1::Vector{<:Integer},: Vector of integers specifying the first site;

b2::Vector{<:Integer},: Vector of integers specifying the second site;

Latlib.write_opsum_to_toml! — Method

Writes into a TOML file all the operators in the OpSum. The file is written in the following format:

    Interactions = [
        ["HB", "J1", [1, 2]],
        ["HB", "J2", [2, 3]],
        ...
    ]

Arguments

opsum::OpSum: OpSum object containing the operators;

filename::String: Name of the file to write to;

Keyword arguments

index_zero::Bool=false: If true, the indices of the sites are written starting from 0 instead of 1;